newversion v7(final)/EdgeSolution/modules/modbusclientmodule/requirements.txt
azure-iot-device
pyModbusTCP
schedule